Polishing PaC form components

+++ api/api-pipelines-as-code-internal-v1/src/main/java/com/thoughtworks/go/apiv1/pipelinesascodeinternal/PipelinesAsCodeInternalControllerV1.java
@@ -39,12 +39,13 @@ import com.thoughtworks.go.domain.materials.Modification;
 import com.thoughtworks.go.plugin.access.configrepo.ConfigFileList;
 import com.thoughtworks.go.plugin.access.configrepo.ExportedConfig;
 import com.thoughtworks.go.server.service.*;
-import com.thoughtworks.go.server.service.plugins.builder.DefaultPluginInfoFinder;
 import com.thoughtworks.go.spark.spring.SparkSpringController;
 import com.thoughtworks.go.util.FileUtil;
 import com.thoughtworks.go.util.SystemEnvironment;
 import org.apache.commons.io.FileUtils;
 import org.apache.commons.lang3.StringUtils;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Component;
@@ -54,27 +55,26 @@ import spark.Response;
 
 import java.io.File;
 import java.util.ArrayList;
-import java.util.HashMap;
+import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.*;
 import java.util.function.Consumer;
 
 import static com.thoughtworks.go.api.util.HaltApiResponses.haltBecauseOfReason;
-import static com.thoughtworks.go.plugin.domain.common.PluginConstants.CONFIG_REPO_EXTENSION;
 import static com.thoughtworks.go.spark.Routes.PaC.*;
 import static java.lang.String.format;
 import static spark.Spark.*;
 
 @Component
 public class PipelinesAsCodeInternalControllerV1 extends ApiController implements SparkSpringController {
+    private static final Logger LOGGER = LoggerFactory.getLogger(PipelinesAsCodeInternalControllerV1.class);
 
     private final ApiAuthenticationHelper apiAuthenticationHelper;
     private final PasswordDeserializer passwordDeserializer;
     private final GoConfigService goConfigService;
     private final GoConfigPluginService pluginService;
     private final PipelineConfigService pipelineService;
-    private DefaultPluginInfoFinder defaultPluginInfoFinder;
     private MaterialService materialService;
     private final MaterialConfigConverter materialConfigConverter;
     private final SubprocessExecutionContext subprocessExecutionContext;
@@ -88,7 +88,6 @@ public class PipelinesAsCodeInternalControllerV1 extends ApiController implement
             GoConfigService goConfigService,
             GoConfigPluginService pluginService,
             PipelineConfigService pipelineService,
-            DefaultPluginInfoFinder defaultPluginInfoFinder,
             MaterialService materialService,
             MaterialConfigConverter materialConfigConverter,
             SubprocessExecutionContext subprocessExecutionContext,
@@ -100,7 +99,6 @@ public class PipelinesAsCodeInternalControllerV1 extends ApiController implement
         this.goConfigService = goConfigService;
         this.pluginService = pluginService;
         this.pipelineService = pipelineService;
-        this.defaultPluginInfoFinder = defaultPluginInfoFinder;
         this.materialService = materialService;
         this.materialConfigConverter = materialConfigConverter;
         this.subprocessExecutionContext = subprocessExecutionContext;
@@ -126,7 +124,9 @@ public class PipelinesAsCodeInternalControllerV1 extends ApiController implement
     }
 
     String configFiles(Request req, Response res) {
-        File folder = FileUtil.createTempFolder();
+        ConfigRepoPlugin repoPlugin = pluginFromRequest(req);
+
+        File folder = null;
 
         try {
             JsonReader jsonReader = GsonTransformer.getInstance().jsonReaderFrom(req.body());
@@ -151,50 +151,31 @@ public class PipelinesAsCodeInternalControllerV1 extends ApiController implement
                 return MessageJson.create("Material is already being used as a config repository");
             }
 
-            Map<String, ConfigFileList> pacPluginFiles = new HashMap<>();
+            folder = FileUtil.createTempFolder();
             checkoutFromMaterialConfig(materialConfig, folder);
-            defaultPluginInfoFinder.pluginDisplayNameToPluginId(CONFIG_REPO_EXTENSION).forEach((k, pluginId) -> {
-                ConfigRepoPlugin crPlugin = (ConfigRepoPlugin) pluginService.partialConfigProviderFor(pluginId);
-                pacPluginFiles.put(pluginId, crPlugin.getConfigFiles(folder, new ArrayList<>()));
-            });
 
+            final Map<String, ConfigFileList> pacPluginFiles = Collections.singletonMap(repoPlugin.id(), repoPlugin.getConfigFiles(folder, new ArrayList<>()));
             return jsonizeAsTopLevelObject(req, w -> ConfigFileListsRepresenter.toJSON(w, pacPluginFiles));
-        } catch (Exception ex) {
-            res.status(HttpStatus.INTERNAL_SERVER_ERROR.value());
-            return MessageJson.create(ex.getMessage());
-        } finally {
-            FileUtils.deleteQuietly(folder);
-        }
-    }
-
-    protected void validateMaterial(MaterialConfig materialConfig) {
-        PipelineConfigSaveValidationContext vctx = PipelineConfigSaveValidationContext.forChain(false, null, goConfigService.getCurrentConfig(), materialConfig);
-        ((ScmMaterialConfig) materialConfig).validateConcreteScmMaterial(vctx);
-    }
-
-    protected void checkoutFromMaterialConfig(MaterialConfig materialConfig, File folder) throws ExecutionException, InterruptedException {
-        Material material = materialConfigConverter.toMaterial(materialConfig);
-        subprocessExecutionContext.setGitShallowClone(true);
-        ExecutorService executor = Executors.newSingleThreadExecutor();
-        Future future = executor.submit(() -> {
-            List<Modification> modifications = materialService.latestModification(material, folder, subprocessExecutionContext);
-            materialService.checkout(material, folder, Modification.latestRevision(modifications), subprocessExecutionContext);
-        });
-        try {
-            future.get(systemEnvironment.getPacCloneTimeout(), TimeUnit.MILLISECONDS);
         } catch (TimeoutException e) {
-            future.cancel(true);
-            throw new RuntimeException("Request couldn't be completed because we failed to clone the repo within 2 minutes");
+            res.status(HttpStatus.PAYLOAD_TOO_LARGE.value());
+            return MessageJson.create("Aborted check because cloning the SCM repository took too long");
+        } catch (ExecutionException e) {
+            res.status(HttpStatus.INTERNAL_SERVER_ERROR.value());
+            return MessageJson.create(e.getCause().getMessage()); // unwrap these exceptions thrown by the future
+        } catch (Exception e) {
+            res.status(HttpStatus.INTERNAL_SERVER_ERROR.value());
+            return MessageJson.create(e.getMessage());
+        } finally {
+            if (null != folder) {
+                FileUtils.deleteQuietly(folder);
+            }
         }
     }
 
     String preview(Request req, Response res) {
-        String pluginId = requiredParam(req, ":plugin_id");
+        ConfigRepoPlugin repoPlugin = pluginFromRequest(req);
+        String pluginId = repoPlugin.id();
         String groupName = requiredQueryParam(req, "group");
-
-        if (!pluginService.isConfigRepoPlugin(pluginId)) {
-            throw haltBecauseOfReason("Plugin `%s` is not a Pipelines-as-Code plugin.", pluginId);
-        }
 
         if (!pluginService.supportsPipelineExport(pluginId)) {
             throw haltBecauseOfReason("Plugin `%s` does not support pipeline config export.", pluginId);
@@ -212,7 +193,6 @@ public class PipelinesAsCodeInternalControllerV1 extends ApiController implement
             return MessageJson.create(format("Please fix the validation errors for pipeline %s.", pipeline.name()), jsonWriter(pipeline));
         }
 
-        ConfigRepoPlugin repoPlugin = plugin(pluginId);
         String etag = repoPlugin.etagForExport(pipeline, groupName);
 
         if (fresh(req, etag)) {
@@ -225,6 +205,30 @@ public class PipelinesAsCodeInternalControllerV1 extends ApiController implement
             res.header("Content-Type", export.getContentType());
             res.header("Content-Disposition", format("attachment; filename=\"%s\"", export.getFilename()));
             return export.getContent();
+        }
+    }
+
+    protected void validateMaterial(MaterialConfig materialConfig) {
+        PipelineConfigSaveValidationContext vctx = PipelineConfigSaveValidationContext.forChain(false, null, goConfigService.getCurrentConfig(), materialConfig);
+        ((ScmMaterialConfig) materialConfig).validateConcreteScmMaterial(vctx);
+    }
+
+    protected void checkoutFromMaterialConfig(MaterialConfig materialConfig, File folder) throws ExecutionException, InterruptedException, TimeoutException {
+        Material material = materialConfigConverter.toMaterial(materialConfig);
+        long timeout = systemEnvironment.getPacCloneTimeout();
+
+        subprocessExecutionContext.setGitShallowClone(true);
+        ExecutorService executor = Executors.newSingleThreadExecutor();
+        Future future = executor.submit(() -> {
+            List<Modification> modifications = materialService.latestModification(material, folder, subprocessExecutionContext);
+            materialService.checkout(material, folder, Modification.latestRevision(modifications), subprocessExecutionContext);
+        });
+        try {
+            future.get(timeout, TimeUnit.MILLISECONDS);
+        } catch (TimeoutException e) {
+            LOGGER.debug(format("Failed to clone material %s in %d ms", material.getDescription(), timeout), e);
+            future.cancel(true);
+            throw e;
         }
     }
 
@@ -256,7 +260,13 @@ public class PipelinesAsCodeInternalControllerV1 extends ApiController implement
         return writer -> PipelineConfigRepresenter.toJSON(writer, pipelineConfig);
     }
 
-    private ConfigRepoPlugin plugin(String pluginId) {
+    private ConfigRepoPlugin pluginFromRequest(Request req) {
+        String pluginId = requiredParam(req, ":plugin_id");
+
+        if (!pluginService.isConfigRepoPlugin(pluginId)) {
+            throw haltBecauseOfReason("Plugin `%s` is not a Pipelines-as-Code plugin.", pluginId);
+        }
+
         return (ConfigRepoPlugin) pluginService.partialConfigProviderFor(pluginId);
     }
